backend/requirements.txt
Flask
flask-cors
gunicorn
pandas
scikit-learn
python-dotenv
pytz
requests
numpy
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.9
Werkzeug
